Move predict related types to predict.ts
