--- docs/reference/main/generative-ai.imagegenerationpredictresponse.md
@@ -1,20 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@google/generative-ai](./generative-ai.md) &gt; [ImageGenerationPredictResponse](./generative-ai.imagegenerationpredictresponse.md)
-
-## ImageGenerationPredictResponse interface
-
-Response message for \[PredictionService.Predict\]. This is an internal class. Please do not depend on it.
-
-**Signature:**
-
-```typescript
-export interface ImageGenerationPredictResponse 
-```
-
-## Properties
-
-|  Property | Modifiers | Type | Description |
-|  --- | --- | --- | --- |
-|  [predictions?](./generative-ai.imagegenerationpredictresponse.predictions.md) |  | [ImageGenerationPredictResponseImageData](./generative-ai.imagegenerationpredictresponseimagedata.md)<!-- -->\[\] | _(Optional)_ The outputs of the prediction call. |
-
