--- docs/reference/main/generative-ai.imagegenerationpredictresponse.predictions.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@google/generative-ai](./generative-ai.md) &gt; [ImageGenerationPredictResponse](./generative-ai.imagegenerationpredictresponse.md) &gt; [predictions](./generative-ai.imagegenerationpredictresponse.predictions.md)
-
-## ImageGenerationPredictResponse.predictions property
-
-The outputs of the prediction call.
-
-**Signature:**
-
-```typescript
-predictions?: ImageGenerationPredictResponseImageData[];
-```
