--- docs/reference/main/generative-ai.imagegenerationpredictresponseimagedata.bytesbase64encoded.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@google/generative-ai](./generative-ai.md) &gt; [ImageGenerationPredictResponseImageData](./generative-ai.imagegenerationpredictresponseimagedata.md) &gt; [bytesBase64Encoded](./generative-ai.imagegenerationpredictresponseimagedata.bytesbase64encoded.md)
-
-## ImageGenerationPredictResponseImageData.bytesBase64Encoded property
-
-**Signature:**
-
-```typescript
-bytesBase64Encoded: string;
-```
