--- docs/reference/main/generative-ai.imagegenerationpredictresponseimagedata.md
@@ -1,21 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@google/generative-ai](./generative-ai.md) &gt; [ImageGenerationPredictResponseImageData](./generative-ai.imagegenerationpredictresponseimagedata.md)
-
-## ImageGenerationPredictResponseImageData interface
-
-Each image data for response of \[PredictionService.Predict\]. This is an internal class. Please do not depend on it.
-
-**Signature:**
-
-```typescript
-export interface ImageGenerationPredictResponseImageData 
-```
-
-## Properties
-
-|  Property | Modifiers | Type | Description |
-|  --- | --- | --- | --- |
-|  [bytesBase64Encoded](./generative-ai.imagegenerationpredictresponseimagedata.bytesbase64encoded.md) |  | string |  |
-|  [mimeType](./generative-ai.imagegenerationpredictresponseimagedata.mimetype.md) |  | string |  |
-
