--- docs/reference/main/generative-ai.imagegenerationpredictresponseimagedata.mimetype.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@google/generative-ai](./generative-ai.md) &gt; [ImageGenerationPredictResponseImageData](./generative-ai.imagegenerationpredictresponseimagedata.md) &gt; [mimeType](./generative-ai.imagegenerationpredictresponseimagedata.mimetype.md)
-
-## ImageGenerationPredictResponseImageData.mimeType property
-
-**Signature:**
-
-```typescript
-mimeType: string;
-```
