--- docs/reference/main/generative-ai.predictrequest.instances.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@google/generative-ai](./generative-ai.md) &gt; [PredictRequest](./generative-ai.predictrequest.md) &gt; [instances](./generative-ai.predictrequest.instances.md)
-
-## PredictRequest.instances property
-
-The instances that are the input to the prediction call.
-
-**Signature:**
-
-```typescript
-instances?: PredictServiceValueType[];
-```
