--- docs/reference/main/generative-ai.predictrequest.md
@@ -1,22 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@google/generative-ai](./generative-ai.md) &gt; [PredictRequest](./generative-ai.predictrequest.md)
-
-## PredictRequest interface
-
-Request message for \[PredictionService.Predict\]\[\]. This is an internal class. Please do not depend on it.
-
-**Signature:**
-
-```typescript
-export interface PredictRequest 
-```
-
-## Properties
-
-|  Property | Modifiers | Type | Description |
-|  --- | --- | --- | --- |
-|  [instances?](./generative-ai.predictrequest.instances.md) |  | PredictServiceValueType\[\] | _(Optional)_ The instances that are the input to the prediction call. |
-|  [model?](./generative-ai.predictrequest.model.md) |  | string | _(Optional)_ The name of the model for prediction. |
-|  [parameters?](./generative-ai.predictrequest.parameters.md) |  | PredictServiceValueType | _(Optional)_ The parameters that govern the prediction call. |
-
