--- docs/reference/main/generative-ai.predictrequest.model.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@google/generative-ai](./generative-ai.md) &gt; [PredictRequest](./generative-ai.predictrequest.md) &gt; [model](./generative-ai.predictrequest.model.md)
-
-## PredictRequest.model property
-
-The name of the model for prediction.
-
-**Signature:**
-
-```typescript
-model?: string;
-```
