--- docs/reference/main/generative-ai.predictrequest.parameters.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@google/generative-ai](./generative-ai.md) &gt; [PredictRequest](./generative-ai.predictrequest.md) &gt; [parameters](./generative-ai.predictrequest.parameters.md)
-
-## PredictRequest.parameters property
-
-The parameters that govern the prediction call.
-
-**Signature:**
-
-```typescript
-parameters?: PredictServiceValueType;
-```
